move get_output
