+++ src/model/openthaigpt_pretraining_model/gptj/gptj_model_xformers.py
@@ -1,14 +1,33 @@
 import torch
-from transformers.models.gptj.modeling_gptj import (
-    GPTJAttention,
-    GPTJModel,
-)
-from transformers import AutoTokenizer
 import xformers.ops as xops
+from xformers.components.attention.core import scaled_query_key_softmax, bmm
 
-_attn_orig = GPTJAttention._attn
 
-device = "cuda" if torch.cuda.is_available() else "cpu"
+def _attn_xformers_cpu(
+    self,
+    query,
+    key,
+    value,
+    attention_mask=None,
+    head_mask=None,
+):
+    # compute causal mask from causal mask buffer
+    query_length, key_length = query.size(-2), key.size(-2)
+    causal_mask = self.bias[:, :, key_length - query_length : key_length, :key_length]
+    attention_mask = torch.where(causal_mask, 0.0, -torch.inf)
+    # Keep the attention weights computation in fp32 to avoid overflow issues
+    query = query.to(torch.float32)
+    key = key.to(torch.float32)
+    attn_weights = scaled_query_key_softmax(query, key, attention_mask)
+    attn_weights = self.attn_dropout(attn_weights)
+
+    # Mask heads if we want to
+    if head_mask is not None:
+        attn_weights = attn_weights * head_mask
+    # attn_output = bmm(attn_weights, value)
+    attn_output = bmm(attn_weights, value)
+
+    return attn_output, attn_weights
 
 
 def _attn_xformers(
@@ -34,19 +53,3 @@ def _attn_xformers(
     )
 
     return attn_output, None
-
-
-def get_output(pretrained_name, use_xformers, input_text):
-    model = GPTJModel.from_pretrained(pretrained_name).to(device)
-    if use_xformers:
-        print("Use xFormers")
-        GPTJAttention._attn = _attn_xformers
-    else:
-        print("Use original")
-        GPTJAttention._attn = _attn_orig
-
-    tokenizer = AutoTokenizer.from_pretrained(pretrained_name)
-    inputs = tokenizer(input_text, return_tensors="pt").to(device)
-    outputs = model(inputs.input_ids)
-    last_hidden_states = outputs.last_hidden_state
-    return last_hidden_states
